Extract IdMap into own module
